requirements.txt
anyio==4.13.0
certifi==2026.4.22
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.13
lxml==6.1.0
PyMuPDF==1.27.2.2
python-docx==1.2.0
python-dotenv==1.2.2
python-telegram-bot==21.6
typing_extensions==4.15.0
